requirements.txt
streamlit==1.40.0
pydeck>=0.8.0
pandas>=2.0.0
numpy>=1.24.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
watchdog>=6.0.0